FIX: refactor code
